+++ src/batch/src/executor/s3_file_scan.rs
@@ -15,8 +15,7 @@
 use futures_async_stream::try_stream;
 use futures_util::stream::StreamExt;
 use risingwave_common::catalog::{Field, Schema};
-use risingwave_connector::source::filesystem::opendal_source::opendal_reader::read_parquet_file;
-use risingwave_connector::source::iceberg::new_s3_operator;
+use risingwave_connector::source::iceberg::{new_s3_operator, read_parquet_file};
 use risingwave_pb::batch_plan::file_scan_node;
 use risingwave_pb::batch_plan::file_scan_node::StorageType;
 use risingwave_pb::batch_plan::plan_node::NodeBody;
