+++ src/connector/src/source/iceberg/parquet_file_reader.rs
@@ -13,26 +13,35 @@
 // limitations under the License.
 
 use std::collections::HashMap;
+use std::future::IntoFuture;
 use std::ops::Range;
+use std::pin::Pin;
 use std::sync::Arc;
 
 use anyhow::anyhow;
 use bytes::Bytes;
 use futures::future::BoxFuture;
-use futures::{FutureExt, TryFutureExt};
+use futures::{FutureExt, Stream, TryFutureExt};
 use iceberg::io::{
     FileIOBuilder, FileMetadata, FileRead, S3_ACCESS_KEY_ID, S3_REGION, S3_SECRET_ACCESS_KEY,
 };
 use iceberg::{Error, ErrorKind};
+use itertools::Itertools;
 use opendal::layers::{LoggingLayer, RetryLayer};
 use opendal::services::S3;
 use opendal::Operator;
 use parquet::arrow::async_reader::AsyncFileReader;
-use parquet::arrow::ParquetRecordBatchStreamBuilder;
-use parquet::file::metadata::{ParquetMetaData, ParquetMetaDataReader};
+use parquet::arrow::{parquet_to_arrow_schema, ParquetRecordBatchStreamBuilder, ProjectionMask};
+use parquet::file::metadata::{FileMetaData, ParquetMetaData, ParquetMetaDataReader};
+use risingwave_common::array::arrow::IcebergArrowConvert;
+use risingwave_common::array::StreamChunk;
+use risingwave_common::catalog::ColumnId;
+use risingwave_common::util::tokio_util::compat::FuturesAsyncReadCompatExt;
 use url::Url;
 
 use crate::error::ConnectorResult;
+use crate::parser::ParquetParser;
+use crate::source::{Column, SourceColumnDesc};
 
 pub struct ParquetFileReader<R: FileRead> {
     meta: FileMetadata,
@@ -168,4 +177,144 @@ pub async fn list_s3_directory(
             format!("Invalid s3 url: {}, should start with {}", dir, prefix),
         ))?
     }
+}
+
+/// Extracts valid column indices from a Parquet file schema based on the user's requested schema.
+///
+/// This function is used for column pruning of Parquet files. It calculates the intersection
+/// between the columns in the currently read Parquet file and the schema provided by the user.
+/// This is useful for reading a `RecordBatch` with the appropriate `ProjectionMask`, ensuring that
+/// only the necessary columns are read.
+///
+/// # Parameters
+/// - `columns`: A vector of `Column` representing the user's requested schema.
+/// - `metadata`: A reference to `FileMetaData` containing the schema and metadata of the Parquet file.
+///
+/// # Returns
+/// - A `ConnectorResult<Vec<usize>>`, which contains the indices of the valid columns in the
+///   Parquet file schema that match the requested schema. If an error occurs during processing,
+///   it returns an appropriate error.
+pub fn extract_valid_column_indices(
+    columns: Option<Vec<Column>>,
+    metadata: &FileMetaData,
+) -> ConnectorResult<Vec<usize>> {
+    match columns {
+        Some(rw_columns) => {
+            let parquet_column_names = metadata
+                .schema_descr()
+                .columns()
+                .iter()
+                .map(|c| c.name())
+                .collect_vec();
+
+            let converted_arrow_schema =
+                parquet_to_arrow_schema(metadata.schema_descr(), metadata.key_value_metadata())
+                    .map_err(anyhow::Error::from)?;
+
+            let valid_column_indices: Vec<usize> = rw_columns
+                .iter()
+                .filter_map(|column| {
+                    parquet_column_names
+                        .iter()
+                        .position(|&name| name == column.name)
+                        .and_then(|pos| {
+                            let arrow_field = IcebergArrowConvert
+                                .to_arrow_field(&column.name, &column.data_type)
+                                .ok()?;
+                            if &arrow_field == converted_arrow_schema.field(pos) {
+                                Some(pos)
+                            } else {
+                                None
+                            }
+                        })
+                })
+                .collect();
+            Ok(valid_column_indices)
+        }
+        None => Ok(vec![]),
+    }
+}
+
+/// Reads a specified Parquet file and converts its content into a stream of chunks.
+pub async fn read_parquet_file(
+    op: Operator,
+    file_name: String,
+    rw_columns: Option<Vec<Column>>,
+    parser_columns: Option<Vec<SourceColumnDesc>>,
+    batch_size: usize,
+    offset: usize,
+) -> ConnectorResult<
+    Pin<Box<dyn Stream<Item = Result<StreamChunk, crate::error::ConnectorError>> + Send>>,
+> {
+    let mut reader: tokio_util::compat::Compat<opendal::FuturesAsyncReader> = op
+        .reader_with(&file_name)
+        .into_future() // Unlike `rustc`, `try_stream` seems require manual `into_future`.
+        .await?
+        .into_futures_async_read(..)
+        .await?
+        .compat();
+    let parquet_metadata = reader.get_metadata().await.map_err(anyhow::Error::from)?;
+
+    let file_metadata = parquet_metadata.file_metadata();
+    let column_indices = extract_valid_column_indices(rw_columns, file_metadata)?;
+    let projection_mask = ProjectionMask::leaves(file_metadata.schema_descr(), column_indices);
+    // For the Parquet format, we directly convert from a record batch to a stream chunk.
+    // Therefore, the offset of the Parquet file represents the current position in terms of the number of rows read from the file.
+    let record_batch_stream = ParquetRecordBatchStreamBuilder::new(reader)
+        .await?
+        .with_batch_size(batch_size)
+        .with_projection(projection_mask)
+        .with_offset(offset)
+        .build()?;
+    let converted_arrow_schema = parquet_to_arrow_schema(
+        file_metadata.schema_descr(),
+        file_metadata.key_value_metadata(),
+    )
+    .map_err(anyhow::Error::from)?;
+    let columns = match parser_columns {
+        Some(columns) => columns,
+        None => converted_arrow_schema
+            .fields
+            .iter()
+            .enumerate()
+            .map(|(index, field_ref)| {
+                let data_type = IcebergArrowConvert.type_from_field(field_ref).unwrap();
+                SourceColumnDesc::simple(
+                    field_ref.name().clone(),
+                    data_type,
+                    ColumnId::new(index as i32),
+                )
+            })
+            .collect(),
+    };
+
+    let parquet_parser = ParquetParser::new(columns, file_name, offset)?;
+    let msg_stream: Pin<
+        Box<dyn Stream<Item = Result<StreamChunk, crate::error::ConnectorError>> + Send>,
+    > = parquet_parser.into_stream(record_batch_stream);
+    Ok(msg_stream)
+}
+
+pub async fn get_parquet_fields(
+    op: Operator,
+    file_name: String,
+) -> ConnectorResult<risingwave_common::array::arrow::arrow_schema_udf::Fields> {
+    let mut reader: tokio_util::compat::Compat<opendal::FuturesAsyncReader> = op
+        .reader_with(&file_name)
+        .into_future() // Unlike `rustc`, `try_stream` seems require manual `into_future`.
+        .await?
+        .into_futures_async_read(..)
+        .await?
+        .compat();
+    let parquet_metadata = reader.get_metadata().await.map_err(anyhow::Error::from)?;
+
+    let file_metadata = parquet_metadata.file_metadata();
+    let converted_arrow_schema = parquet_to_arrow_schema(
+        file_metadata.schema_descr(),
+        file_metadata.key_value_metadata(),
+    )
+    .map_err(anyhow::Error::from)?;
+    let fields: risingwave_common::array::arrow::arrow_schema_udf::Fields =
+        converted_arrow_schema.fields;
+    Ok(fields)
 }
